добавление файлов darwin и linux
